requirements.txt
crawl4ai>=0.4.247
python-dotenv>=1.0.0
pydantic>=2.10.0
aiohttp>=3.8.0
google-generativeai>=0.3.0
pandas>=2.0.0
